fix(JSAM-18):change helper to services folder
